back/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.12.14
click==8.1.7
distro==1.9.0
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
mangum==0.19.0
openai==1.57.4
pydantic==2.10.3
pydantic_core==2.27.1
python-dotenv==1.0.1
setuptools==75.1.0
sniffio==1.3.1
starlette==0.41.3
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0
wheel==0.44.0
langchain == 0.3.12
langchain-core == 0.3.25
langchain-openai == 0.2.12
langchain-text-splitters == 0.3.3
langchain-upstage == 0.4.0
langchain_pinecone == 0.2.0
pinecone == 5.3.1
langgraph == 0.2.59
google-search-results
langchain-community